Move MustExec and Exec helpers into apitest package
